Refactor input parsing code into separate class
